# Don't use order in product query by default